[processMedia] fix redis undefined error
Currently identical image / media article fast-forward mechanism will trigger the following error:
Error: node_redis: The SET command contains a invalid argument type of "undefined".
This is because we return choosingArticle() in processImage & processMedia, but the return signature is not compatible.
This commit reorganizes the return value of choosingArticle() in processImage() and processMedia() so that it works.